infernce log field updated

+++ ai_project/models.py
@@ -46,7 +46,7 @@ class InferenceLog(BaseModel):
     model = models.ForeignKey(AIModel, on_delete=models.DO_NOTHING, null=True)
     input_params = models.TextField(default="", blank=True)
     output_details = models.TextField(default="", blank=True)
-    total_inference_time = models.IntegerField(default=0)
+    total_inference_time = models.FloatField(default=0)
     total_credits_used = models.FloatField(default=0)
     status = models.CharField(max_length=255, default="")   # success, failed, in_progress
